// src/types/Brand.ts
export interface ShapeSwiperItem {
  title: string;
  subTitle: string;
  desc: string;
  imgSrcPc: string;
  imgSrcMo: string;
  wPc: number;
  wMo: number;
  hPc: number;
  hMo: number;
}

export type ShapeItem = {
  id: string;
  imgPc: string;
  imgMo: string;
  wPc: number;
  hPc: number;
  wMo: number;
  hMo: number;
};